refactor: move AuthProvider.tsx and api/ out of app/
